// src/modules/user/domain/enum/domainException.enum.ts
export enum DomainExectionCode {
	DEFAULT_DOMAIN_EXCEPTION = 'DEFAULT_DOMAIN_EXCEPTION',
	USER_NAME_REQUIRED = 'USER_NAME_REQUIRED',
	USER_LASTNAME_REQUIRED = 'USER_LASTNAME_REQUIRED',
	USER_LOUNGE_REQUIRED = 'USER_LOUNGE_REQUIRED',
	USER_LOUNGE_INVALID = 'USER_LOUNGE_INVALID',
	USER_CODE_REQUIRED = 'USER_CODE_REQUIRED',
	USER_CODE_LENGTH_INVALID = 'USER_CODE_LENGTH_INVALID',
	USER_GUID_INVALID = 'USER_GUID_INVALID',
	USER_NOT_FOUND = 'USER_NOT_FOUND',
	LOUNGE_NOT_FOUND = 'LOUNGE_NOT_FOUND',
}